Switch imports to work for new db function locations. Add logs_controller
