+++ shuttle/driver_logs/__init__.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from flask import render_template
 from flask_classy import FlaskView, route, request
 
-from shuttle.schedules.shuttle_schedules_controller import ScheduleController
+from shuttle.driver_logs.logs_controller import DriverLogsController
 from shuttle.shuttle_controller import ShuttleController
 
 
@@ -12,12 +12,12 @@ class DriverLogsView(FlaskView):
 
     def __init__(self):
         self.sc = ShuttleController()
-        self.ssc = ScheduleController()
+        self.dlc = DriverLogsController()
 
     @route('/driver-logs')
     def shuttle_logs(self):
         self.sc.check_roles_and_route(['Administrator'])
-        grabbed_logs = self.ssc.grab_logs()
+        grabbed_logs = self.dlc.grab_logs()
         now = datetime.now()
         date = now.strftime('%b-%d-%Y')
         logs = grabbed_logs[0]
@@ -32,7 +32,7 @@ class DriverLogsView(FlaskView):
             sort = json_data['sort']
         else:
             sort = ''
-        selected_logs = self.ssc.grab_selected_logs(date, sort)
+        selected_logs = self.dlc.grab_selected_logs(date, sort)
         shuttle_logs = selected_logs[0]
         break_logs = selected_logs[1]
         return render_template('driver_logs/load_logs.html', **locals())
