+++ shuttle/request_shuttle/__init__.py
@@ -1,7 +1,7 @@
 from flask import render_template
 from flask_classy import FlaskView, route, request
 
-from shuttle.db import db_functions as db
+from shuttle.db.db_tables import shuttle_request_logs_functions as request_db
 from shuttle.schedules.google_sheets_controller import SheetsController
 from shuttle.schedules.shuttle_schedules_controller import ScheduleController
 from shuttle.shuttle_controller import ShuttleController
@@ -19,16 +19,16 @@ class RequestShuttleView(FlaskView):
     def request(self):
         self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
         locations = self.shc.grab_locations()
-        active_requests = db.number_active_requests()
+        active_requests = request_db.number_active_requests()
         if active_requests['requested-pick-up']:
-            position_in_waitlist = db.get_position_in_waitlist()[0]['rownumber']
+            position_in_waitlist = request_db.get_position_in_waitlist()[0]['rownumber']
         return render_template('request_shuttle/request_shuttle.html', **locals())
 
     @route('/send-request', methods=['GET', 'POST'])
     def send_shuttle_request_path(self):
         self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
         json_data = request.get_json()
-        response = db.commit_shuttle_request(json_data['pick-up-location'], json_data['drop-off-location'])
+        response = request_db.commit_shuttle_request(json_data['pick-up-location'], json_data['drop-off-location'])
         if response == 'success':
             self.sc.set_alert('success', 'Your request has been submitted')
         elif response == 'same location':
@@ -43,7 +43,7 @@ class RequestShuttleView(FlaskView):
     @route('/delete-request')
     def delete_request(self):
         self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
-        request_to_delete = db.delete_current_request()
+        request_to_delete = request_db.delete_current_request()
         if request_to_delete == 'success':
             self.sc.set_alert('success', 'Your request has been deleted')
         else:
