+++ shuttle/schedules/__init__.py
@@ -1,7 +1,7 @@
 from flask import render_template
 from flask_classy import FlaskView, route
 
-from shuttle.db import db_functions as db
+from shuttle.db.db_tables import shuttle_schedule_functions as schedule_db
 from shuttle.schedules.google_sheets_controller import SheetsController
 from shuttle.schedules.shuttle_schedules_controller import ScheduleController
 from shuttle.shuttle_controller import ShuttleController
@@ -31,7 +31,7 @@ class SchedulesView(FlaskView):
         self.sc.check_roles_and_route(['Administrator'])
         spreadsheet = self.shc.grab_schedule()
         locations = self.shc.grab_locations()
-        sent = db.commit_schedule(spreadsheet, locations)
+        sent = schedule_db.commit_schedule(spreadsheet, locations)
         if sent == "success":
             self.sc.set_alert('success', 'The schedule has been submitted')
         elif sent == 'no match':
